requirements.txt
fastapi
uvicorn
jinja2
pyyaml
supabase
dotenv
email-validator